fix: checked prop validation warning
Ensures that imageIsSelected function returns a boolean value when
evaluating selected state for Radio checked prop. The issue that was
causing the warning was the fact that the expression (("") && ("" ===
"foo")) returns empty string instead of a boolean value. This is only
the case when working with an empty string as the input value to
imageIsSelected function.